Dedupe some boilerplate
Add a macro to make it slightly easier to link a new day's code into
main.rs. I was hoping to have the macro generate the `mod dayXX;`
statement but couldn't figure out how to do that without causing
rust-analyzer to think the referenced file was not included in the build
graph, which would turn off IDE features.