Finish day 6, part 2
The original implementation for part 1 turned out to be super
inefficient, since it iterated over every single fish on every day. Now
we just keep track of the number of fish there are for each number of
days before spawning. The runtime is now linear in terms of the number
of days, as opposed to exponential, and uses effectively constant space.